make the examples more easy to port
