--- examples/auto_test/requirements.txt
@@ -1,1 +1,0 @@
-tensorflow
