+++ examples/cifar10/pic/pic2.h
@@ -1,4 +1,4 @@
-const unsigned char pic[32*32*3]={\
+static const unsigned char pic[32*32*3]={\
  62,116, 47,  60,111, 51,  64,113, 59,  68,119, 62,  74,127, 64,  79,132, 69,  82,133, 76,  89,134, 85,  94,136, 89,  97,136, 91, 104,143, 98, 100,142, 95, 100,145, 89, 102,146, 87, 106,145, 90, 113,146,101, 107,156, 88, 110,157, 95, 113,154,103, 118,157,112, 120,156,112, 120,157,111, 127,163,116, 126,162,115, 128,164,118, 130,166,120, 133,169,122, 136,174,122, 140,178,126, 142,178,131, 141,175,134, 140,172,137, 
  67,118, 51,  66,114, 56,  71,117, 64,  74,123, 67,  79,130, 68,  82,135, 72,  85,135, 77,  92,138, 85,  98,142, 89, 103,144, 93, 101,143, 95, 100,143, 92, 105,153, 94, 104,151, 88, 110,154, 95, 113,152,100, 113,159, 93, 114,158, 99, 116,158,105, 125,162,116, 127,164,118, 127,165,117, 133,170,120, 133,170,120, 134,170,123, 137,173,127, 139,175,129, 142,179,129, 144,181,131, 148,182,135, 147,179,138, 146,176,141, 
  72,116, 55,  73,115, 60,  76,119, 68,  81,126, 70,  84,133, 71,  88,137, 75,  90,138, 79,  96,141, 84, 102,145, 88, 104,146, 91, 108,149, 98, 116,159,108, 115,163,105, 107,156, 94, 104,152, 93, 109,155,102, 115,157, 96, 116,157, 99, 119,159,104, 128,165,115, 131,167,120, 131,168,118, 134,173,118, 134,173,118, 136,175,123, 138,176,128, 141,177,133, 144,179,135, 147,180,135, 148,181,136, 150,180,139, 149,177,141,
