+++ examples/cifar10/pic/pic3.h
@@ -1,4 +1,4 @@
-const unsigned char pic[32*32*3]={\
+static const unsigned char pic[32*32*3]={\
  67, 41, 27,  61, 38, 22,  62, 46, 23,  86, 77, 44,  86, 86, 50,  91, 96, 65, 103,110, 89, 134,140,123, 148,149,133, 135,127,114, 146,128,117, 192,168,148, 159,135, 99, 169,144,102, 175,145,118, 181,144,136, 208,176,165, 184,150,137, 165,136,121, 134,119,100, 128,128,110, 219,223,212, 204,196,196, 180,163,167, 193,179,180, 144,142,132, 129,132,117,  73, 64, 55,  91, 66, 70, 109, 79, 90, 114,102,108,  95,100, 99, 
  54, 40, 18,  60, 46, 24,  67, 53, 30,  86, 78, 48,  97, 94, 63,  87, 88, 62,  95, 97, 77, 117,119,100, 112,112, 94, 114,107, 92, 109, 97, 85, 149,134,118, 157,144,118, 150,139,111, 149,133,121, 113, 93, 98, 123,104,101, 111, 88, 86, 137,115,110, 129,119,109, 158,159,149, 242,246,240, 235,229,230, 195,180,184, 172,160,160, 145,144,134, 128,132,120, 131,127,122, 121,101,106,  94, 70, 82, 131,120,128, 123,127,128, 
  57, 53, 25,  67, 60, 35,  63, 56, 31,  70, 63, 36,  79, 75, 47,  76, 74, 50,  80, 79, 58,  80, 79, 59,  98, 94, 75, 127,122,107, 109,102, 93, 133,130,122, 143,147,136, 125,132,125, 123,129,136, 117,121,140, 138,137,141, 105, 95,101, 118,104,108,  86, 78, 79, 106,107,105, 222,224,224, 204,199,201, 183,169,171, 126,114,110, 112,111, 97, 100,103, 88, 134,131,123, 127,113,117,  99, 82, 93, 115,105,111, 116,116,116,
