+++ examples/cifar10/pic/pic4.h
@@ -1,4 +1,4 @@
-const unsigned char pic[32*32*3]={\
+static const unsigned char pic[32*32*3]={\
  55, 72, 63,  47, 68, 59,  66, 84, 77,  86, 95, 92,  84, 89, 88,  89,101,103,  90,111,113,  84,109,111,  91,110,115,  89,101,107,  83, 98,101,  90,103,105,  72, 77, 80,  85, 84, 86,  89, 94, 93,  82, 95, 93,  84,112,106,  73, 99, 93,  59, 83, 73,  71, 88, 79,  65, 78, 70,  82, 89, 82,  87, 91, 86,  75, 76, 72,  83, 84, 80,  81, 84, 75,  85, 89, 78,  88, 90, 78,  77, 78, 68,  77, 75, 67,  85, 80, 77,  80, 75, 74, 
  71, 84, 76,  57, 72, 64,  74, 88, 82,  89, 96, 93,  89, 91, 91,  91,100,103,  91,111,112,  87,110,112,  93,112,115,  96,111,114,  80, 95, 98,  83, 96, 98,  77, 85, 85,  83, 88, 87,  78, 86, 85,  71, 86, 82,  64, 90, 84,  68, 91, 83,  74, 93, 84,  75, 91, 80,  72, 81, 71,  87, 91, 85,  86, 87, 83,  93, 94, 90,  81, 82, 78,  76, 79, 70,  78, 82, 71,  85, 87, 75,  83, 85, 73,  82, 80, 72,  88, 85, 81,  85, 80, 79, 
  70, 77, 70,  62, 69, 64,  75, 80, 78,  86, 87, 85, 102,102,102, 101,109,109,  89,104,106,  88,107,110,  86,105,108,  90,108,109,  80, 98, 99,  72, 88, 87,  71, 84, 82,  90,101, 98,  80, 94, 90,  77, 95, 88,  72, 93, 85,  67, 85, 78,  73, 86, 78,  75, 84, 74,  92, 97, 88,  90, 92, 86,  92, 93, 89,  85, 86, 84,  78, 79, 75,  75, 77, 71,  74, 78, 67,  78, 80, 68,  80, 82, 70,  81, 79, 71,  83, 80, 76,  78, 74, 73,
