+++ examples/cifar10/pic/pic5.h
@@ -1,4 +1,4 @@
-const unsigned char pic[32*32*3]={\
+static const unsigned char pic[32*32*3]={\
 105,110, 41, 111,109, 44, 121,110, 48, 131,116, 53, 131,120, 52, 129,123, 50, 127,127, 51, 126,123, 49, 132,124, 54, 139,124, 55, 147,132, 60, 141,131, 54, 134,128, 51, 123,114, 46, 132,120, 62, 135,118, 69, 147,133, 51, 144,136, 59, 138,129, 61, 138,123, 60, 144,125, 64, 146,131, 68, 141,134, 65, 135,125, 55, 143,127, 58, 149,129, 58, 150,136, 64, 133,124, 50, 132,117, 48, 125,104, 42, 132,111, 56, 133,118, 69, 
 105,106, 44, 111,108, 47, 123,112, 52, 127,115, 51, 124,115, 47, 120,119, 45, 117,121, 45, 119,121, 45, 128,123, 48, 131,121, 44, 129,118, 38, 136,126, 42, 136,125, 43, 139,126, 50, 133,116, 49, 144,123, 62, 140,124, 65, 134,123, 66, 141,131, 83, 159,143,100, 156,137, 94, 150,135, 86, 131,122, 65, 137,126, 66, 146,125, 64, 150,126, 66, 149,131, 68, 136,122, 56, 137,119, 56, 143,119, 61, 146,123, 73, 132,114, 67, 
 103,106, 44, 108,106, 42, 118,112, 47, 120,111, 43, 114,108, 37, 110,109, 35, 109,111, 35, 119,119, 43, 131,127, 48, 136,127, 47, 129,118, 34, 139,125, 42, 140,128, 46, 137,123, 47, 139,123, 54, 146,126, 61, 139,121, 84, 170,157,125, 205,193,169, 213,199,177, 218,198,173, 201,184,151, 140,129, 85, 136,122, 70, 154,130, 78, 154,126, 72, 146,123, 67, 143,125, 66, 141,121, 63, 147,122, 66, 148,124, 72, 129,110, 59,
