+++ examples/cifar10/pic/pic7.h
@@ -1,4 +1,4 @@
-const unsigned char pic[32*32*3]={\
+static const unsigned char pic[32*32*3]={\
  46, 48, 56,  63, 66, 71,  84, 88, 89,  96,103, 96,  98,108, 92, 113,125,107, 128,137,127, 200,208,208, 242,248,253, 235,244,248, 240,248,248, 226,234,233, 219,225,224, 218,223,221, 211,215,209, 197,203,192, 186,195,192, 169,177,176, 160,166,165, 174,180,179, 203,208,207, 200,202,202, 204,206,206, 204,204,204, 196,196,196, 206,206,206, 209,210,208, 210,211,209, 157,158,154,  60, 64, 59,  49, 53, 48,  40, 44, 39, 
   9, 11, 11,  17, 19, 19,  14, 18, 13,  35, 42, 27,  55, 64, 43,  93,105, 83, 133,144,128, 175,184,181, 178,186,186, 168,178,178, 161,172,169, 146,157,154, 136,145,142, 141,148,143, 131,141,129, 127,137,121, 122,130,129, 127,133,132, 126,132,131, 166,171,170, 227,232,231, 233,235,235, 239,241,241, 241,241,241, 247,247,247, 241,241,241, 246,247,245, 253,254,252, 165,166,162,  27, 31, 26,  21, 25, 20,  21, 25, 20, 
  35, 38, 23,  36, 39, 24,  31, 35, 16,  55, 64, 37, 101,112, 79, 113,126, 94, 127,139,117, 129,138,128, 141,150,147, 130,141,139, 124,138,134, 132,146,142, 132,146,142, 151,162,159, 191,205,194, 200,215,201, 208,214,213, 219,225,224, 187,193,192, 162,167,166, 241,246,245, 247,249,249, 247,249,249, 254,254,254, 253,253,253, 249,249,249, 251,252,250, 250,251,249, 162,163,159,  28, 32, 27,  19, 23, 18,  18, 22, 17,
