+++ examples/cifar10/pic/pic8.h
@@ -1,4 +1,4 @@
-const unsigned char pic[32*32*3]={\
+static const unsigned char pic[32*32*3]={\
  96,120,142,  94,122,139,  93,124,139,  95,122,142,  89,113,135,  99,125,142, 106,132,144, 108,130,142, 108,123,139, 109,120,134, 103,119,131,  96,120,138,  91,121,150,  92,120,155, 107,129,154, 112,131,139, 102,128,135, 102,128,140, 107,132,148, 107,131,149,  99,124,140,  97,123,139,  96,122,138,  90,115,135,  86,110,134,  87,110,136,  88,111,137,  86,110,132,  84,109,129,  85,109,131,  87,111,135,  87,108,135, 
  90,117,143,  85,116,137,  85,118,138,  87,118,143,  82,112,139,  88,120,143,  91,124,143,  92,124,143,  94,118,140, 105,127,145, 105,130,146,  83,116,136,  78,117,149,  79,117,152,  95,129,153, 101,131,142, 101,126,136,  98,121,136, 100,122,140, 101,125,143,  99,123,141,  96,121,137,  93,119,135,  89,115,132,  84,108,130,  90,114,138,  90,114,138,  84,109,129,  81,107,124,  88,114,131,  92,116,138,  89,113,137, 
  84,114,143,  81,115,139,  83,120,142,  84,119,145,  78,112,142,  79,113,142,  75,113,137,  79,117,141,  83,116,142,  99,127,151, 107,139,158,  83,122,144,  83,127,158,  82,124,159,  93,134,157, 107,143,153, 111,132,147, 105,125,142, 102,121,142, 102,123,145, 101,122,143,  93,118,134,  90,115,131,  90,116,133,  90,115,135,  87,111,133,  86,111,131,  84,113,128,  85,112,126,  87,114,128,  89,115,131,  87,112,132,
