+++ examples/cifar10/tm_example_cifar10.h
@@ -1,0 +1,17 @@
+#ifndef TM_EXAMPLE_CIFAR10_H
+#define TM_EXAMPLE_CIFAR10_H
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+#include "../tm_examples.h"
+
+    int tm_example_cifar10();
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*TM_EXAMPLE_CIFAR10_H*/
