+++ examples/mbnet/pic128.h
@@ -10,7 +10,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-const unsigned char pic[128*128*3]={\
+static const unsigned char pic[128*128*3]={\
  35, 72,  0,  36, 73,  0,  37, 74,  5,  37, 74,  5,  37, 73,  6,  36, 70, 11,  33, 64,  9,  30, 59,  7,  27, 53,  5,  27, 51,  5,  29, 52,  7,  31, 53,  9,  29, 53,  9,  29, 56, 11,  33, 63, 16,  28, 59, 12,  26, 59,  7,  23, 54,  3,  20, 47,  3,  23, 44,  9,  25, 41, 14,  29, 42, 22,  31, 42, 27,  32, 41, 23,  31, 43, 18,  28, 51, 18,  28, 61, 21,  34, 67, 29,  39, 74, 34,  44, 84, 39,  43, 89, 41,  44, 94, 41,  46, 98, 45,  49,100, 49,  44, 97, 44,  41, 94, 40,  41, 96, 40,  41, 97, 40,  36, 96, 32,  34, 91, 23,  33, 84, 16,  37, 82, 12,  37, 76,  6,  35, 74,  4,  31, 74,  5,  32, 77,  7,  35, 81, 16,  32, 79, 26,  34, 81, 40,  40, 82, 46,  44, 86, 45,  39, 86, 37,  35, 86, 33,  31, 87, 30,  35, 88, 34,  42, 83, 33,  46, 86, 35,  50, 88, 35,  56, 93, 41,  61, 94, 42,  59, 87, 43,  53, 79, 40,  46, 66, 33,  38, 52, 26,  33, 42, 20,  28, 34, 21,  16, 26, 22,  15, 22, 16,  15, 19, 14,  20, 22, 14,  24, 25, 16,  28, 23, 15,  29, 25, 13,  28, 27, 10,  29, 31, 15,  30, 32, 25,  25, 31, 26,  22, 34, 23,  26, 37, 27,  28, 39, 28,  27, 39, 27,  30, 41, 31,  36, 40, 31,  41, 43, 33,  43, 46, 32,  48, 51, 34,  53, 57, 38,  56, 59, 40,  59, 61, 45,  63, 64, 47,  65, 65, 47,  66, 65, 47,  71, 69, 49,  67, 69, 54,  65, 69, 55,  68, 71, 57,  72, 75, 61,  73, 76, 63,  78, 79, 66,  86, 84, 72,  86, 84, 72,  83, 81, 69,  83, 81, 69,  83, 81, 69,  86, 84, 72,  84, 82, 70,  78, 76, 64,  79, 77, 65,  77, 75, 63,  77, 73, 61,  79, 74, 63,  74, 76, 63,  70, 75, 61,  77, 75, 63,  89, 76, 68,  82, 79, 67,  83, 86, 72,  84, 86, 72,  84, 84, 71,  85, 83, 70,  83, 83, 70,  77, 80, 66,  79, 81, 68,  79, 79, 67,  82, 79, 68,  84, 79, 68,  84, 80, 73,  84, 81, 74, 
  38, 64,  3,  36, 62,  4,  32, 59,  2,  34, 59,  6,  36, 60,  9,  34, 56, 11,  32, 52, 10,  30, 50, 10,  26, 43,  6,  30, 42,  8,  33, 44, 13,  30, 43, 12,  26, 41,  9,  27, 44, 11,  28, 46, 13,  24, 44, 11,  23, 48,  8,  21, 46,  6,  19, 39,  6,  16, 33,  7,  22, 35, 15,  28, 40, 22,  31, 44, 26,  35, 45, 26,  33, 47, 22,  31, 54, 21,  31, 65, 23,  36, 70, 30,  38, 74, 32,  46, 88, 42,  50, 97, 47,  56,108, 53,  54,109, 53,  53,110, 56,  48,105, 50,  45,103, 47,  42,100, 44,  35, 94, 37,  36, 91, 32,  32, 85, 22,  31, 81, 16,  32, 76,  7,  32, 72,  1,  31, 72,  1,  31, 74,  2,  32, 77,  4,  33, 80, 10,  33, 80, 22,  34, 82, 35,  39, 83, 35,  39, 82, 30,  38, 84, 31,  40, 90, 31,  37, 91, 30,  39, 90, 28,  43, 86, 27,  50, 91, 32,  54, 94, 35,  59, 97, 38,  63, 98, 38,  55, 88, 36,  47, 77, 32,  41, 65, 29,  34, 51, 26,  29, 41, 25,  26, 32, 20,  22, 21,  8,  22, 19,  6,  25, 19,  5,  30, 23,  5,  34, 25,  7,  29, 29,  9,  27, 33,  8,  26, 34,  6,  25, 35,  9,  25, 36, 18,  29, 37, 22,  32, 33, 13,  32, 33, 13,  35, 36, 16,  38, 39, 20,  38, 39, 20,  38, 43, 22,  42, 50, 24,  46, 53, 27,  49, 57, 27,  54, 63, 30,  57, 64, 37,  61, 65, 50,  62, 66, 50,  62, 66, 50,  63, 66, 51,  68, 70, 54,  67, 69, 55,  66, 67, 55,  69, 71, 58,  74, 76, 63,  75, 77, 64,  80, 81, 69,  87, 85, 73,  87, 85, 73,  86, 84, 72,  86, 84, 72,  86, 84, 72,  87, 85, 73,  88, 86, 74,  84, 82, 70,  84, 82, 70,  80, 78, 66,  77, 75, 64,  77, 76, 67,  71, 75, 65,  73, 77, 67,  73, 71, 63,  73, 63, 57,  66, 62, 55,  59, 59, 50,  63, 61, 53,  81, 79, 70,  88, 85, 77,  84, 84, 75,  81, 84, 70,  79, 81, 68,  80, 79, 67,  84, 81, 70,  86, 81, 71,  89, 85, 77,  88, 85, 78, 
  40, 54, 13,  37, 51, 11,  32, 46,  8,  33, 45, 10,  31, 44, 11,  31, 42, 13,  31, 40, 13,  31, 38, 14,  29, 34, 14,  30, 30, 13,  28, 28, 13,  27, 27, 13,  28, 30, 14,  29, 34, 17,  26, 34, 16,  24, 33, 15,  22, 36, 13,  19, 35, 11,  18, 32, 11,  17, 30, 13,  20, 30, 17,  24, 37, 22,  28, 45, 24,  34, 49, 24,  34, 51, 22,  31, 58, 21,  32, 69, 25,  34, 72, 28,  43, 82, 36,  50, 94, 45,  54,104, 50,  58,112, 55,  58,118, 57,  51,116, 55,  47,109, 50,  42,101, 45,  38, 96, 42,  31, 87, 36,  31, 79, 29,  29, 75, 20,  32, 77, 16,  32, 74,  8,  32, 73,  1,  31, 73,  0,  34, 78,  3,  37, 84,  6,  35, 84,  7,  34, 83, 15,  34, 84, 25,  39, 86, 20,  38, 83, 18,  38, 84, 24,  41, 87, 26,  39, 88, 21,  41, 90, 18,  46, 92, 21,  55, 97, 29,  60,101, 34,  60, 98, 35,  55, 91, 28,  49, 86, 29,  43, 77, 28,  37, 63, 25,  29, 49, 24,  22, 37, 22,  18, 27, 18,  18, 20, 10,  20, 20, 10,  23, 21,  9,  28, 24, 10,  31, 26, 12,  29, 31, 18,  28, 35, 18,  30, 39, 17,  27, 39, 19,  24, 37, 25,  28, 38, 27,  34, 39, 21,  34, 39, 21,  34, 38, 20,  36, 41, 23,  37, 41, 23,  42, 47, 28,  44, 51, 30,  50, 57, 34,  52, 59, 34,  56, 65, 36,  61, 67, 43,  64, 65, 51,  62, 66, 50,  60, 65, 51,  60, 67, 55,  62, 70, 58,  70, 72, 60,  71, 70, 58,  73, 72, 60,  76, 76, 64,  78, 79, 66,  82, 82, 70,  88, 86, 74,  88, 86, 74,  86, 84, 72,  85, 83, 71,  84, 82, 70,  85, 83, 71,  87, 85, 73,  85, 83, 71,  85, 83, 71,  83, 81, 69,  80, 79, 69,  75, 78, 71,  75, 78, 71,  58, 58, 52,  37, 35, 30,  36, 29, 25,  41, 33, 31,  30, 22, 21,  25, 20, 17,  48, 43, 40,  80, 78, 75,  87, 88, 81,  83, 86, 72,  80, 82, 69,  81, 80, 68,  85, 82, 70,  88, 83, 72,  90, 86, 78,  89, 86, 79,
