+++ examples/mbnet/pic96.h
@@ -10,7 +10,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-const unsigned char pic[96*96*3]={\
+static const unsigned char pic[96*96*3]={\
  36, 71,  0,  36, 71,  2,  37, 72,  5,  37, 72,  7,  34, 66, 11,  30, 58,  7,  27, 51,  4,  29, 51,  7,  31, 52,  9,  28, 52,  8,  31, 58, 14,  28, 59, 12,  24, 57,  6,  20, 49,  3,  20, 42,  7,  25, 40, 15,  30, 43, 25,  34, 42, 25,  31, 44, 19,  28, 56, 19,  33, 66, 28,  40, 75, 35,  45, 87, 41,  46, 95, 44,  49,100, 48,  48,100, 49,  43, 97, 42,  42, 98, 41,  39, 97, 37,  35, 92, 25,  33, 83, 15,  36, 78,  8,  34, 73,  3,  30, 73,  3,  32, 79,  9,  33, 79, 24,  35, 81, 42,  43, 83, 46,  40, 86, 38,  34, 87, 32,  33, 89, 32,  41, 85, 33,  46, 86, 34,  53, 91, 38,  62, 96, 43,  58, 87, 42,  50, 73, 36,  39, 54, 27,  33, 41, 22,  22, 29, 20,  16, 20, 14,  18, 20, 11,  24, 24, 12,  28, 24, 13,  28, 27,  9,  27, 30,  9,  29, 33, 22,  25, 32, 23,  25, 34, 21,  29, 38, 25,  30, 39, 25,  33, 40, 28,  40, 43, 30,  44, 48, 30,  50, 55, 34,  55, 60, 38,  60, 62, 46,  64, 65, 48,  67, 66, 47,  71, 69, 50,  65, 68, 54,  67, 70, 56,  72, 75, 62,  75, 78, 65,  85, 84, 72,  86, 83, 72,  84, 82, 70,  84, 82, 70,  87, 85, 73,  82, 80, 68,  79, 77, 65,  77, 75, 63,  78, 73, 62,  74, 76, 63,  73, 78, 64,  86, 77, 68,  84, 79, 68,  81, 84, 70,  86, 86, 73,  87, 85, 73,  82, 82, 70,  78, 81, 68,  80, 79, 67,  83, 79, 68,  85, 81, 71,  85, 82, 76, 
  40, 59,  8,  34, 54,  6,  33, 52,  7,  34, 52, 11,  32, 48, 12,  31, 44, 12,  28, 38, 10,  31, 36, 13,  30, 35, 12,  28, 36, 12,  28, 41, 15,  25, 39, 14,  22, 43, 10,  19, 39,  9,  16, 32, 10,  21, 33, 17,  28, 41, 25,  35, 47, 26,  34, 50, 21,  31, 62, 22,  34, 70, 28,  41, 79, 35,  50, 96, 46,  57,110, 54,  56,115, 57,  51,112, 54,  45,105, 48,  38, 98, 43,  33, 89, 34,  31, 81, 22,  31, 78, 14,  32, 73,  3,  31, 71,  0,  32, 76,  2,  33, 80,  5,  32, 80, 16,  35, 82, 31,  39, 83, 26,  38, 84, 26,  40, 90, 29,  38, 91, 24,  43, 89, 23,  53, 95, 30,  59, 99, 36,  60, 96, 34,  51, 87, 32,  42, 72, 28,  32, 52, 26,  24, 38, 25,  20, 24, 13,  20, 17,  5,  25, 20,  5,  32, 23,  6,  30, 29, 12,  25, 35, 12,  25, 37, 11,  22, 37, 20,  29, 37, 23,  34, 35, 15,  35, 36, 17,  38, 40, 20,  39, 42, 22,  42, 50, 26,  47, 56, 29,  51, 62, 30,  58, 67, 38,  62, 66, 52,  60, 65, 51,  60, 66, 53,  66, 70, 57,  68, 70, 57,  70, 71, 58,  75, 76, 64,  78, 79, 67,  86, 85, 73,  87, 85, 73,  86, 84, 72,  85, 83, 71,  86, 84, 72,  86, 84, 72,  84, 82, 70,  81, 78, 66,  77, 77, 68,  75, 80, 72,  64, 68, 60,  52, 47, 42,  52, 45, 41,  40, 37, 32,  53, 49, 45,  85, 81, 78,  86, 87, 78,  80, 83, 69,  80, 79, 67,  85, 81, 69,  89, 84, 75,  89, 86, 80, 
  37, 46, 14,  35, 43, 14,  31, 38, 13,  27, 35, 12,  29, 35, 14,  29, 31, 15,  25, 24, 12,  23, 18,  9,  23, 17, 11,  24, 22, 14,  22, 24, 14,  20, 24, 14,  19, 24, 11,  16, 23,  9,  18, 27, 14,  19, 28, 18,  23, 41, 21,  32, 49, 22,  32, 52, 18,  30, 65, 20,  35, 76, 28,  47, 89, 40,  57,107, 52,  62,119, 59,  57,120, 55,  46,111, 47,  37, 95, 39,  33, 85, 36,  28, 73, 29,  30, 70, 24,  30, 71, 13,  30, 71,  2,  35, 77,  2,  42, 88,  8,  44, 92,  7,  42, 93, 15,  39, 91, 20,  37, 86,  9,  35, 80, 18,  39, 81, 22,  43, 87, 15,  48, 95, 17,  58,101, 29,  64,104, 37,  56, 92, 31,  48, 81, 30,  39, 66, 25,  31, 49, 22,  22, 32, 16,  15, 23, 17,  14, 21, 17,  16, 21, 14,  23, 25, 17,  27, 25, 18,  28, 25, 15,  28, 30, 17,  29, 32, 27,  26, 34, 29,  27, 38, 26,  25, 36, 24,  25, 37, 25,  33, 40, 29,  47, 46, 35,  49, 51, 37,  57, 58, 42,  60, 61, 41,  66, 64, 44,  67, 68, 52,  61, 67, 55,  61, 68, 57,  72, 70, 59,  74, 71, 60,  79, 75, 64,  83, 80, 68,  89, 87, 75,  88, 86, 75,  87, 85, 74,  86, 84, 72,  87, 86, 73,  84, 82, 70,  84, 82, 70,  86, 84, 72,  87, 87, 77,  66, 62, 55,  46, 35, 30,  39, 30, 23,  34, 22, 19,  29, 16, 16,  15,  8,  6,  58, 55, 52,  88, 91, 82,  82, 85, 71,  83, 83, 71,  86, 82, 70,  90, 85, 76,  92, 89, 82,
