+++ examples/mbnet/tm_example_mbnet.h
@@ -1,0 +1,17 @@
+#ifndef TM_EXAMPLE_MBNET_H
+#define TM_EXAMPLE_MBNET_H
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+#include "../tm_examples.h"
+
+    int tm_example_mbnet();
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*TM_EXAMPLE_MBNET_H*/
