+++ examples/mnist/tm_example_mnist.h
@@ -1,0 +1,17 @@
+#ifndef TM_EXAMPLE_MNIST_H
+#define TM_EXAMPLE_MNIST_H
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+#include "../tm_examples.h"
+
+    int tm_example_mnist();
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*TM_EXAMPLE_MNIST_H*/
