+++ examples/vww/pic/img2c.py
@@ -30,7 +30,7 @@ if __name__ == '__main__':
     img = Image.open(img_name)
     img = np.array(img)
     fw = open("pic.h", "w")
-    fw.writelines("const unsigned char pic[%d*%d*%d]={\\\n"%(img.shape[0],img.shape[1],img.shape[2]))
+    fw.writelines("static const unsigned char pic[%d*%d*%d]={\\\n"%(img.shape[0],img.shape[1],img.shape[2]))
     for y in range(img.shape[0]):
         for x in range(img.shape[1]):
             for c in range(img.shape[2]):
