+++ examples/vww/pic/pic_noperson1.h
@@ -1,4 +1,4 @@
-const unsigned char pic[96*96*3]={\
+static const unsigned char pic[96*96*3]={\
 156,166,175, 159,169,178, 161,171,180, 162,172,181, 162,172,181, 162,172,181, 163,173,182, 165,175,184, 167,177,186, 167,177,186, 168,178,187, 167,177,186, 167,177,186, 166,176,185, 165,175,184, 164,174,183, 164,171,190, 163,170,189, 163,170,189, 162,169,188, 161,168,187, 159,167,186, 158,167,184, 157,166,183, 158,167,184, 156,168,184, 157,169,185, 158,170,184, 158,172,185, 159,173,186, 159,173,186, 160,174,187, 150,168,188, 147,162,181, 145,158,175, 149,158,173, 156,164,177, 161,169,182, 158,170,186, 154,169,188, 154,174,198, 138,162,188, 144,168,194, 143,165,189, 150,167,187, 146,158,174,  77, 83, 95,   8, 11, 20,  18, 19, 23,  71, 71, 79, 143,147,158, 111,118,134, 147,155,174, 152,165,184, 148,165,183, 153,172,187, 157,176,190, 153,173,184, 150,168,180, 150,166,181, 151,163,185, 153,160,189, 153,157,194, 153,156,197, 147,168,187, 151,170,187, 155,174,191, 160,177,193, 166,179,195, 166,178,192, 163,173,185, 159,167,178, 173,181,192, 170,176,188, 168,174,188, 170,178,191, 174,181,197, 169,178,195, 159,168,185, 149,158,175, 127,145,167, 125,143,165, 134,150,173, 152,169,189, 165,180,201, 165,177,199, 159,170,192, 155,166,188, 152,162,187, 152,162,189, 146,157,185, 134,148,177, 126,141,172, 124,138,173, 121,137,171, 118,134,170, 
 155,165,174, 157,167,176, 160,170,179, 160,170,179, 160,170,179, 160,170,179, 162,172,181, 164,174,183, 165,175,184, 165,175,184, 166,176,185, 166,176,185, 165,175,184, 164,174,183, 163,173,182, 163,173,182, 161,168,187, 161,168,187, 160,167,186, 160,167,186, 159,166,185, 157,166,183, 157,166,183, 157,166,183, 155,164,181, 154,166,182, 156,168,182, 157,169,183, 158,172,185, 159,173,186, 160,174,187, 161,175,188, 152,170,192, 148,165,185, 147,160,177, 150,159,174, 156,164,177, 158,167,182, 156,168,184, 152,167,186, 149,169,193, 135,159,185, 142,166,192, 138,160,184, 144,161,181, 142,154,170,  77, 83, 95,  11, 14, 23,  15, 16, 18,  66, 69, 74, 142,146,157, 113,120,136, 146,157,177, 150,165,186, 146,162,185, 147,165,185, 145,167,181, 148,167,181, 144,163,177, 139,155,171, 139,151,173, 145,154,183, 149,155,191, 149,153,191, 148,169,188, 149,168,185, 148,167,184, 151,168,184, 159,172,188, 166,178,192, 174,184,196, 178,186,199, 182,190,201, 181,189,202, 180,188,201, 180,187,203, 178,185,201, 174,183,200, 169,178,195, 163,174,192, 149,167,189, 147,163,186, 151,168,188, 163,178,199, 171,183,205, 166,178,200, 161,172,194, 159,170,192, 154,164,189, 151,163,189, 143,154,182, 128,142,171, 120,135,168, 118,134,168, 117,133,169, 113,131,167, 
 153,163,172, 155,165,174, 157,167,176, 157,167,176, 157,167,176, 157,167,176, 159,169,178, 161,171,180, 162,172,181, 163,173,182, 163,173,182, 163,173,182, 163,173,182, 162,172,181, 161,171,180, 161,171,180, 158,165,183, 158,165,183, 158,165,183, 158,165,183, 158,165,183, 156,165,182, 156,165,182, 156,165,182, 152,164,180, 153,165,179, 155,167,181, 157,169,183, 158,172,185, 159,173,186, 160,174,187, 160,174,187, 152,170,192, 150,167,187, 149,162,179, 152,161,176, 156,164,177, 157,166,181, 154,166,182, 149,166,186, 144,164,188, 132,156,182, 140,164,190, 132,154,178, 136,153,173, 135,147,163,  76, 82, 94,  13, 17, 28,  13, 15, 14,  61, 65, 68, 141,145,154, 114,121,137, 143,154,176, 146,160,186, 140,157,183, 133,153,177, 132,153,174, 140,161,178, 140,159,176, 129,146,164, 128,143,166, 141,152,180, 148,157,190, 147,153,189, 149,170,189, 151,169,189, 151,170,187, 153,170,186, 160,173,189, 170,182,196, 182,191,206, 191,199,212, 182,190,203, 183,191,204, 183,190,206, 179,186,202, 173,182,199, 169,177,196, 167,178,196, 168,179,197, 164,181,201, 162,179,199, 164,179,200, 169,182,201, 168,181,200, 164,175,195, 160,171,191, 160,171,193, 153,163,188, 147,159,185, 133,147,174, 117,132,163, 108,124,157, 108,124,160, 107,125,163, 104,124,161,
