+++ examples/vww/pic/pic_noperson2.h
@@ -1,4 +1,4 @@
-const unsigned char pic[96*96*3]={\
+static const unsigned char pic[96*96*3]={\
  97, 98, 92,  97, 98, 92,  97, 98, 92,  98, 99, 93,  98, 99, 93,  98, 99, 93,  99,100, 94,  99,100, 94,  99,100, 94,  99,100, 94,  99,100, 94, 100,101, 95, 100,101, 95, 100,101, 95, 101,102, 96, 101,102, 96,  99,100, 95,  99,100, 94,  99,100, 94,  99,100, 94,  99,100, 94,  99,100, 94, 100,100, 92, 100,100, 92,  98, 98, 90, 102, 99, 92, 101, 98, 89, 100, 97, 88, 101, 98, 89, 103, 99, 90,  97, 93, 82,  88, 84, 73,  74, 75, 61,  91, 92, 78, 101,102, 88,  97, 97, 85,  94, 94, 82,  98, 98, 86,  97, 97, 85,  90, 90, 80,  95, 95, 85,  94, 94, 84,  94, 94, 86,  94, 94, 86,  94, 94, 86,  94, 94, 86,  93, 93, 85,  93, 93, 85,  92, 92, 84,  92, 92, 84,  91, 91, 83,  90, 90, 82,  89, 89, 81,  89, 89, 81,  88, 88, 80,  88, 88, 80,  88, 88, 80,  88, 88, 80,  87, 87, 79,  86, 86, 78,  86, 86, 78,  85, 85, 77,  84, 84, 76,  84, 84, 76,  84, 84, 72,  84, 84, 72,  84, 84, 72,  83, 83, 71,  83, 83, 71,  82, 82, 70,  82, 82, 70,  82, 82, 70,  81, 81, 69,  81, 81, 69,  81, 81, 69,  80, 80, 68,  80, 80, 68,  79, 79, 67,  79, 79, 67,  79, 79, 67,  80, 77, 70,  80, 77, 70,  80, 77, 70,  79, 76, 69,  79, 76, 69,  78, 75, 68,  78, 75, 68,  78, 75, 68,  79, 76, 69,  79, 76, 69,  80, 77, 70,  82, 79, 72,  85, 82, 75,  86, 83, 76,  84, 81, 74,  82, 79, 72, 
  99,100, 94,  99,100, 94,  99,100, 94, 100,101, 95, 100,101, 95, 100,101, 95, 101,102, 96, 101,102, 96, 100,101, 95, 100,101, 95, 100,101, 95, 101,102, 96, 101,102, 96, 101,102, 96, 102,103, 97, 102,103, 97, 101,102, 96, 101,102, 96, 101,102, 96, 101,102, 96, 101,102, 96, 101,102, 96, 102,102, 94, 102,102, 94, 100,100, 92, 103,100, 91, 102, 99, 90, 101, 98, 89, 103,100, 91, 106,102, 93, 101, 97, 86,  93, 89, 78,  74, 75, 61,  90, 91, 77, 100,101, 87,  97, 97, 85,  94, 94, 82,  98, 98, 86,  98, 98, 86,  93, 93, 83,  95, 95, 85,  95, 95, 85,  95, 95, 87,  95, 95, 87,  94, 94, 86,  94, 94, 86,  94, 94, 86,  93, 93, 85,  92, 92, 84,  92, 92, 84,  91, 91, 83,  91, 91, 83,  90, 90, 82,  89, 89, 81,  88, 88, 80,  88, 88, 80,  88, 88, 80,  88, 88, 80,  87, 87, 79,  87, 87, 79,  86, 86, 78,  85, 85, 77,  84, 84, 76,  84, 84, 76,  84, 84, 72,  84, 84, 72,  84, 84, 72,  83, 83, 71,  83, 83, 71,  82, 82, 70,  82, 82, 70,  82, 82, 70,  81, 81, 69,  81, 81, 69,  81, 81, 69,  80, 80, 68,  80, 80, 68,  79, 79, 67,  79, 79, 67,  79, 79, 67,  80, 77, 70,  80, 77, 70,  80, 77, 70,  79, 76, 69,  79, 76, 69,  78, 75, 68,  78, 75, 68,  78, 75, 68,  79, 76, 69,  79, 76, 69,  80, 77, 70,  82, 79, 72,  85, 82, 75,  86, 83, 76,  84, 81, 74,  82, 79, 72, 
 101,102, 96, 101,102, 96, 102,103, 97, 102,103, 97, 102,103, 97, 103,104, 98, 103,104, 98, 103,104, 98, 101,102, 96, 102,103, 97, 102,103, 97, 102,103, 97, 103,104, 98, 103,104, 98, 103,104, 98, 104,105, 99, 104,105, 99, 104,105, 99, 104,105, 99, 104,105, 99, 104,105, 99, 104,105, 99, 105,105, 97, 105,105, 97, 103,103, 95, 106,103, 94, 105,102, 93, 103,100, 91, 106,103, 94, 110,106, 95, 106,102, 91,  98, 94, 83,  80, 81, 67,  94, 95, 81, 102,103, 89,  99,100, 86,  96, 96, 84,  98, 98, 86,  99, 99, 87,  95, 95, 85,  96, 96, 86,  96, 96, 86,  96, 96, 86,  95, 95, 87,  95, 95, 87,  95, 95, 87,  94, 94, 86,  94, 94, 86,  93, 93, 85,  93, 93, 85,  92, 92, 84,  91, 91, 83,  90, 90, 82,  90, 90, 82,  89, 89, 81,  89, 89, 81,  88, 88, 80,  88, 88, 80,  88, 88, 80,  87, 87, 79,  86, 86, 78,  85, 85, 77,  85, 85, 77,  84, 84, 76,  84, 84, 72,  84, 84, 72,  84, 84, 72,  83, 83, 71,  83, 83, 71,  82, 82, 70,  82, 82, 70,  82, 82, 70,  81, 81, 69,  81, 81, 69,  81, 81, 69,  80, 80, 68,  80, 80, 68,  79, 79, 67,  79, 79, 67,  79, 79, 67,  80, 77, 70,  80, 77, 70,  80, 77, 70,  79, 76, 69,  79, 76, 69,  78, 75, 68,  78, 75, 68,  78, 75, 68,  79, 76, 69,  79, 76, 69,  80, 77, 70,  83, 80, 73,  85, 82, 75,  86, 83, 76,  85, 82, 75,  83, 80, 73,
