+++ examples/vww/pic/pic_noperson3.h
@@ -1,4 +1,4 @@
-const unsigned char pic[96*96*3]={\
+static const unsigned char pic[96*96*3]={\
 147,164,207, 146,163,206, 145,162,205, 146,163,206, 148,167,209, 151,170,210, 151,170,210, 149,171,210, 151,173,212, 151,173,210, 150,172,209, 148,172,208, 147,171,205, 145,169,203, 145,169,203, 144,168,202, 140,169,213, 140,169,213, 139,168,212, 139,168,212, 138,167,211, 137,166,210, 137,166,210, 136,165,209, 135,164,208, 135,164,208, 134,163,207, 132,161,205, 131,160,204, 129,158,202, 128,157,201, 127,156,200, 126,152,201, 125,151,200, 125,151,200, 123,149,198, 122,148,197, 121,147,196, 120,146,195, 120,146,195, 121,147,196, 121,147,196, 121,147,196, 121,147,196, 120,146,195, 120,146,195, 120,146,195, 120,146,195, 124,148,196, 124,148,196, 123,147,195, 122,146,194, 121,145,193, 120,144,192, 119,143,191, 119,143,191, 119,143,191, 119,143,191, 121,145,193, 123,147,195, 124,148,196, 126,150,198, 128,152,200, 128,152,200, 126,150,198, 126,150,198, 126,150,198, 126,150,198, 126,150,198, 125,149,197, 125,149,197, 125,149,197, 122,146,194, 126,150,198, 124,148,196, 118,142,190, 121,145,193, 131,155,203, 132,156,204, 125,149,197, 116,145,189, 116,145,189, 116,145,189, 115,144,188, 115,144,188, 114,143,187, 114,143,187, 114,143,187, 117,146,190, 117,146,190, 117,146,190, 118,147,191, 118,147,191, 119,148,192, 119,148,192, 119,148,192, 
 150,167,210, 148,165,208, 146,165,207, 147,166,208, 149,168,210, 151,170,212, 149,171,210, 148,170,209, 151,173,212, 150,172,211, 148,172,208, 147,171,207, 146,170,206, 145,169,205, 143,169,204, 142,168,203, 140,169,213, 139,168,212, 138,167,211, 138,167,211, 136,165,209, 136,165,209, 135,164,208, 134,163,207, 133,162,206, 133,162,206, 132,161,205, 130,159,203, 129,158,202, 128,157,201, 127,156,200, 126,155,199, 125,151,200, 125,151,200, 124,150,199, 124,150,199, 123,149,198, 122,148,197, 122,148,197, 122,148,197, 122,148,197, 122,148,197, 122,148,197, 122,148,197, 123,149,198, 123,149,198, 123,149,198, 123,149,198, 126,150,198, 126,150,198, 125,149,197, 124,148,196, 124,148,196, 123,147,195, 122,146,194, 122,146,194, 124,148,196, 124,148,196, 125,149,197, 125,149,197, 126,150,198, 127,151,199, 127,151,199, 128,152,200, 127,151,199, 127,151,199, 127,151,199, 127,151,199, 126,150,198, 126,150,198, 126,150,198, 126,150,198, 126,150,198, 124,148,196, 121,145,193, 121,145,193, 127,151,199, 132,156,204, 128,152,200, 121,145,193, 116,145,189, 116,145,189, 116,145,189, 115,144,188, 115,144,188, 115,144,188, 114,143,187, 114,143,187, 115,144,188, 116,145,189, 116,145,189, 116,145,189, 117,146,190, 117,146,190, 118,147,191, 118,147,191, 
 152,171,213, 150,169,211, 149,168,210, 149,168,210, 148,170,211, 149,171,212, 149,171,212, 148,170,211, 148,171,212, 148,172,210, 147,171,209, 145,171,208, 144,170,207, 143,169,206, 142,168,205, 141,167,204, 138,167,211, 138,167,211, 137,166,210, 136,165,209, 134,163,207, 133,162,206, 132,161,205, 131,160,204, 130,159,203, 130,159,203, 129,158,202, 128,157,201, 127,156,200, 126,155,199, 125,154,198, 125,154,198, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 125,152,199, 126,153,200, 126,153,200, 127,154,201, 127,154,201, 128,155,202, 128,155,202, 128,155,202, 127,154,201, 127,154,201, 126,153,200, 125,152,199, 125,152,199, 124,151,198, 128,155,202, 127,154,201, 127,154,201, 127,154,201, 127,154,201, 127,154,201, 127,154,201, 126,153,200, 127,154,201, 127,154,201, 126,153,200, 126,153,200, 125,152,199, 125,152,199, 124,151,198, 124,151,198, 128,155,202, 120,147,194, 117,144,191, 124,151,198, 131,158,205, 128,155,202, 121,148,195, 115,142,189, 116,145,189, 116,145,189, 116,145,189, 116,145,189, 115,144,188, 115,144,188, 115,144,188, 114,143,187, 115,144,188, 115,144,188, 115,144,188, 115,144,188, 116,145,189, 116,145,189, 116,145,189, 116,145,189,
