+++ examples/vww/pic/pic_person1.h
@@ -1,4 +1,4 @@
-const unsigned char pic[96*96*3]={\
+static const unsigned char pic[96*96*3]={\
 155,156,176, 104,105,125, 122,123,143, 155,158,177,  94, 99,118,  75, 80, 99,  73, 79, 95,  77, 83, 99, 133,139,155, 161,164,179, 167,170,185, 165,164,180, 164,161,178, 148,146,160, 145,141,156, 168,161,177, 159,166,174, 160,167,175, 158,165,173,  97,104,112,  71, 78, 86,  79, 86, 94,  74, 81, 89, 126,133,141, 142,149,157, 114,121,129, 142,149,157, 147,154,162,  80, 87, 95,  70, 77, 85,  59, 66, 74,  68, 75, 83,  43, 53, 52,  37, 47, 46,  42, 52, 51,  68, 80, 78, 112,124,122, 145,157,155, 149,161,159, 137,149,147, 137,147,146, 149,159,158, 154,163,162, 144,150,150, 132,136,137, 127,131,132, 122,123,125, 114,115,117,  92, 91, 96,  87, 86, 91,  81, 81, 83,  75, 75, 77,  70, 70, 72,  64, 64, 66,  59, 59, 61,  55, 55, 57,  67, 67, 67,  93, 95, 94, 113,115,114, 118,120,119, 129,131,130, 146,148,145, 142,144,141, 124,126,123,  90, 88, 93,  84, 83, 88,  87, 88, 92, 106,110,113, 119,124,127, 116,126,127, 141,153,151, 185,200,197, 230,245,242, 235,250,247, 241,253,251, 241,252,248, 238,247,244, 240,246,244, 246,252,250, 252,255,255, 252,254,253, 249,251,250, 246,248,247, 247,249,248, 249,251,250, 251,253,252, 249,251,250, 247,249,248, 254,255,255, 254,255,255, 254,255,255, 253,255,254, 253,255,254, 252,254,253, 252,254,253, 252,254,253, 
 139,138,156,  90, 91,109, 101,103,118, 133,136,151,  81, 84, 99,  60, 66, 80,  57, 63, 77,  60, 66, 78, 100,106,118, 129,133,145, 136,138,150, 132,132,144, 130,128,141, 114,110,124, 108,102,116, 127,121,135, 119,123,132, 119,123,132, 118,122,131,  61, 65, 74,  33, 37, 46,  34, 38, 47,  23, 27, 36,  65, 69, 78,  81, 85, 94,  51, 55, 64,  78, 82, 91,  99,103,112,  55, 59, 68,  64, 68, 77,  74, 78, 87,  97,101,110, 165,174,173, 153,162,161, 128,137,136, 101,110,109,  86, 95, 94,  84, 93, 92,  86, 95, 94,  86, 95, 94, 102,111,110, 103,109,109,  98,104,104,  89, 93, 94,  85, 86, 88,  84, 85, 87,  84, 84, 86,  81, 81, 83,  67, 66, 71,  64, 63, 68,  60, 60, 62,  56, 56, 58,  53, 53, 55,  48, 48, 50,  43, 43, 45,  38, 39, 41,  29, 31, 30,  44, 46, 45,  55, 57, 56,  57, 59, 58,  68, 70, 69,  86, 88, 87,  91, 93, 92,  84, 86, 85,  76, 74, 79,  80, 79, 84,  90, 91, 95, 108,112,115, 114,122,124, 112,122,123, 142,154,154, 191,206,203, 217,232,229, 230,245,242, 247,255,255, 247,255,255, 248,255,255, 247,255,253, 246,252,250, 248,252,251, 254,255,255, 253,255,254, 250,252,251, 250,252,251, 253,255,254, 254,255,255, 253,255,254, 252,254,253, 252,254,253, 252,254,253, 251,253,252, 251,253,252, 250,252,251, 250,252,251, 250,252,251, 249,251,250, 
 167,167,179, 128,128,140, 128,128,140, 160,163,172, 120,124,133, 103,107,116,  99,103,112, 101,106,112, 121,126,132, 151,154,161, 159,162,167, 156,157,162, 155,154,160, 140,137,144, 133,128,135, 149,144,151, 130,135,141, 129,134,140, 131,136,142,  86, 91, 97,  59, 64, 70,  55, 60, 66,  42, 47, 53,  71, 76, 82,  97,102,108,  60, 65, 71,  69, 74, 80,  90, 95,101,  49, 54, 60,  50, 55, 61,  57, 62, 68,  80, 85, 91, 122,123,125, 139,140,142, 151,152,154, 146,147,149, 130,131,133, 113,117,118, 104,108,109, 100,101,103,  76, 77, 79,  67, 68, 70,  60, 61, 63,  59, 59, 61,  59, 59, 61,  58, 56, 59,  58, 56, 59,  60, 58, 61,  50, 49, 54,  48, 47, 52,  46, 45, 50,  44, 44, 46,  43, 43, 45,  40, 40, 42,  35, 36, 38,  32, 33, 35,  36, 37, 39,  38, 39, 41,  38, 40, 39,  36, 38, 37,  39, 41, 40,  46, 48, 47,  49, 53, 52,  48, 52, 51,  48, 47, 52,  57, 56, 61,  68, 69, 73,  80, 84, 87,  85, 93, 95,  94,104,105, 139,151,151, 199,213,213, 218,232,232, 228,243,240, 238,253,250, 242,254,252, 239,249,248, 239,248,247, 246,252,252, 251,255,255, 252,255,255, 252,255,255, 250,254,253, 250,254,253, 251,255,254, 252,255,255, 252,255,255, 252,255,255, 251,255,254, 251,255,254, 250,254,253, 250,254,253, 250,254,253, 249,253,252, 249,253,252, 249,253,252,
