+++ examples/vww/pic/pic_person2.h
@@ -1,4 +1,4 @@
-const unsigned char pic[96*96*3]={\
+static const unsigned char pic[96*96*3]={\
  23,  0,  0,  23,  2,  1,  24,  4,  3,  24,  6,  4,  22,  7,  4,  19,  4,  1,  16,  1,  0,  14,  0,  0,  22,  3,  0,  20,  0,  0,  24,  0,  0,  35,  1,  0,  43,  0,  0,  43,  0,  0,  43,  0,  0,  49,  0,  0,  48,  1,  0,  51,  4,  0,  55,  8,  0,  58, 11,  1,  53,  9,  0,  45,  2,  0,  38,  0,  0,  35,  0,  0,  28,  0,  0,  28,  0,  0,  31,  6,  2,  31, 12,  8,  24,  8,  8,  13,  1,  3,  12,  1,  5,  15,  6,  9,  19,  3,  0,  18,  4,  0,  15,  5,  0,  12,  6,  0,   6,  4,  0,   3,  4,  0,   1,  6,  0,   2,  8,  4,   0,  8,  5,   0,  6,  3,   0,  4,  4,   0,  3,  4,   3,  5,  4,   5,  5,  5,   6,  4,  5,   5,  3,  4,  18,  0,  0,  27,  0,  5,  34,  2,  5,  37,  2,  0,  43,  5,  0,  53, 10,  0,  56, 11,  0,  53,  7,  0,  57, 12,  0,  51, 10,  0,  43,  5,  0,  34,  3,  0,  29,  3,  6,  24,  3, 12,  19,  1, 17,  14,  0, 17,  27,  2,  0,  23,  0,  0,  25,  1,  0,  30,  6,  4,  33, 12,  9,  31, 10,  7,  29,  5,  5,  27,  3,  3,  29,  0,  2,  32,  0,  3,  37,  1,  3,  41,  0,  4,  43,  0,  3,  46,  0,  1,  48,  0,  2,  50,  0,  3,  37,  2,  0,  37,  2,  0,  38,  0,  0,  36,  0,  0,  37,  0,  0,  41,  0,  0,  49,  5,  6,  54,  8, 10,  60, 14, 16,  49,  0,  4,  56,  5, 10,  69, 16, 22,  66, 11, 17,  58,  2, 11,  61,  4, 13,  61,  4, 13, 
  24,  3,  2,  25,  4,  3,  26,  6,  5,  27,  9,  9,  24,  8,  8,  22,  6,  6,  19,  4,  1,  19,  1,  0,  23,  4,  0,  25,  0,  0,  32,  1,  0,  45,  8,  0,  52,  9,  0,  56,  8,  0,  62, 10,  0,  69, 15,  5,  67, 15,  4,  66, 16,  5,  68, 19,  5,  69, 20,  6,  64, 16,  2,  56,  8,  0,  44,  0,  0,  38,  0,  0,  35,  0,  0,  31,  0,  0,  32,  3,  0,  33,  8,  4,  27,  8,  4,  22,  4,  4,  19,  5,  5,  23,  9,  9,  29, 10,  0,  29, 11,  0,  29, 12,  2,  25, 12,  3,  18, 11,  1,  13, 10,  3,   9, 10,  4,   9, 11,  8,   3,  7,  6,   1,  5,  4,   0,  3,  2,   1,  3,  2,   3,  3,  3,   5,  3,  4,   7,  3,  4,   7,  1,  3,  19,  0,  6,  27,  1, 10,  35,  6, 10,  39,  8,  5,  50, 13,  4,  62, 20,  6,  67, 22,  3,  64, 17,  0,  61, 14,  0,  57, 12,  0,  51,  9,  0,  45,  8,  0,  39,  8,  6,  34,  5, 10,  28,  2, 11,  23,  0, 12,  36,  6,  4,  33,  3,  1,  31,  3,  0,  34,  6,  5,  35,  9,  8,  35,  9,  8,  31,  7,  7,  29,  3,  4,  29,  3,  4,  33,  3,  5,  35,  3,  4,  38,  2,  4,  39,  0,  3,  41,  0,  1,  44,  0,  2,  45,  0,  2,  36,  3,  0,  36,  3,  0,  35,  2,  0,  34,  0,  0,  34,  0,  0,  38,  0,  0,  44,  4,  4,  49,  7,  8,  58, 14, 15,  47,  1,  4,  48,  0,  3,  52,  3,  7,  49,  0,  3,  48,  0,  1,  53,  0,  6,  55,  2,  8, 
  24,  2,  4,  24,  4,  5,  25,  7,  7,  26,  8,  8,  24,  8,  8,  22,  6,  6,  20,  5,  2,  20,  2,  0,  22,  1,  0,  28,  3,  0,  41,  8,  1,  53, 15,  6,  62, 18,  9,  70, 20,  9,  81, 27, 15,  92, 36, 23,  92, 35, 18,  91, 34, 17,  90, 33, 16,  89, 32, 15,  87, 31, 14,  82, 26, 11,  70, 19,  2,  60, 11,  0,  59, 13,  0,  50,  8,  0,  43,  6,  0,  42,  9,  0,  38,  9,  3,  32,  7,  2,  31,  7,  5,  31, 10,  7,  35,  7,  0,  36,  9,  0,  36, 11,  4,  33, 13,  6,  25, 10,  3,  17,  6,  2,  12,  4,  1,   8,  4,  3,   7,  6,  4,   5,  3,  4,   4,  2,  3,   6,  2,  3,   7,  3,  4,   9,  3,  5,  10,  1,  4,  10,  0,  3,  12,  0,  6,  16,  0,  7,  23,  1,  4,  33,  5,  4,  47, 14,  7,  62, 22, 12,  71, 25, 10,  71, 22,  5,  71, 20,  1,  71, 20,  3,  69, 20,  5,  67, 21,  8,  63, 20, 13,  56, 17, 12,  48, 10,  9,  42,  6,  6,  50, 10, 11,  45,  6,  7,  40,  4,  4,  39,  5,  6,  39,  7,  8,  40, 10, 10,  38, 10,  9,  33,  7,  6,  33,  7,  6,  32,  6,  5,  31,  5,  4,  33,  5,  4,  33,  3,  3,  34,  3,  1,  37,  3,  2,  37,  3,  2,  32,  4,  0,  32,  4,  0,  31,  3,  0,  30,  1,  0,  28,  0,  0,  31,  0,  0,  37,  3,  1,  42,  7,  5,  52, 14, 13,  47,  7,  7,  43,  1,  2,  41,  0,  0,  41,  0,  0,  44,  0,  0,  46,  0,  2,  47,  1,  3,
