+++ examples/vww/pic/pic_person3.h
@@ -1,4 +1,4 @@
-const unsigned char pic[96*96*3]={\
+static const unsigned char pic[96*96*3]={\
  89,127,146,  89,127,146,  89,127,146,  90,128,147,  90,128,147,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  92,130,149,  92,130,149,  92,130,149,  93,131,150,  93,131,150,  94,132,151,  94,132,151,  94,132,151,  93,131,150,  93,131,150,  93,131,150,  94,132,151,  94,132,151,  95,133,152,  95,133,152,  95,133,152,  92,134,148,  92,134,148,  92,134,150,  93,135,151,  93,135,151,  94,136,152,  94,136,152,  94,136,152,  94,135,153,  97,135,154,  98,136,155,  98,136,155,  98,136,155,  98,136,157,  99,137,158, 100,138,159,  99,136,154,  99,136,154,  99,136,154, 100,137,155, 100,137,155, 101,138,156, 101,138,156, 101,138,156, 100,137,155, 100,137,155, 100,137,155, 101,138,156, 101,138,156, 102,139,157, 102,139,157, 102,139,157, 102,138,160, 102,138,160, 102,138,160, 103,139,161, 103,139,161, 104,140,162, 104,140,162, 104,140,162, 104,140,162, 104,140,162, 104,140,162, 105,141,163, 105,141,163, 106,142,164, 106,142,164, 106,142,164, 105,141,163, 105,141,163, 105,141,163, 106,142,164, 106,142,164, 106,142,164, 107,143,165, 107,143,165, 109,145,167, 109,145,167, 109,145,167, 110,146,168, 110,146,168, 111,147,169, 111,147,169, 111,147,169, 
  89,127,146,  89,127,146,  90,128,147,  90,128,147,  90,128,147,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  92,130,149,  92,130,149,  93,131,150,  93,131,150,  93,131,150,  94,132,151,  94,132,151,  94,132,151,  93,131,150,  93,131,150,  94,132,151,  94,132,151,  94,132,151,  95,133,152,  95,133,152,  95,133,152,  92,134,148,  92,134,148,  93,135,149,  93,135,151,  93,135,151,  94,136,152,  94,136,152,  96,135,152,  97,135,154,  97,135,154,  98,136,155,  98,136,155,  97,135,154,  97,135,154,  98,136,155,  99,137,156,  99,136,154,  99,136,154, 100,137,155, 100,137,155, 100,137,155, 101,138,156, 101,138,156, 101,138,156, 100,137,155, 100,137,155, 101,138,156, 101,138,156, 101,138,156, 102,139,157, 102,139,157, 102,139,157, 102,138,160, 102,138,160, 103,139,161, 103,139,161, 103,139,161, 104,140,162, 104,140,162, 104,140,162, 104,140,162, 104,140,162, 105,141,163, 105,141,163, 105,141,163, 106,142,164, 106,142,164, 106,142,164, 105,141,163, 105,141,163, 106,142,164, 106,142,164, 106,142,164, 107,143,165, 107,143,165, 107,143,165, 109,145,167, 109,145,167, 110,146,168, 110,146,168, 110,146,168, 111,147,169, 111,147,169, 111,147,169, 
  89,127,146,  90,128,147,  90,128,147,  90,128,147,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  91,129,148,  92,130,149,  93,131,150,  93,131,150,  93,131,150,  94,132,151,  94,132,151,  94,132,151,  94,132,151,  93,131,150,  94,132,151,  94,132,151,  94,132,151,  95,133,152,  95,133,152,  95,133,152,  95,133,152,  92,134,148,  93,135,149,  93,135,149,  93,135,151,  94,136,152,  96,135,152,  96,135,152,  96,135,152,  98,137,154,  98,137,154,  98,136,155,  97,135,154,  97,135,154,  98,135,154,  98,135,154,  98,135,154, 100,135,154, 101,136,155, 101,136,155, 101,136,155, 102,137,156, 102,137,156, 102,137,156, 102,137,156, 101,136,155, 102,137,156, 102,137,156, 102,137,156, 103,138,157, 103,138,157, 103,138,157, 103,138,157, 103,138,158, 104,139,159, 104,139,159, 104,139,159, 105,140,160, 105,140,160, 105,140,160, 105,140,160, 105,140,160, 106,141,161, 106,141,161, 106,141,161, 107,142,162, 107,142,162, 107,142,162, 107,142,162, 107,142,162, 107,142,162, 107,142,162, 108,143,163, 108,143,163, 108,143,163, 109,144,164, 109,144,164, 110,145,165, 111,146,166, 111,146,166, 111,146,166, 112,147,167, 112,147,167, 112,147,167, 112,147,167,
