+++ examples/vww/tm_example_vww.h
@@ -1,0 +1,17 @@
+#ifndef TM_EXAMPLE_VWW_H
+#define TM_EXAMPLE_VWW_H
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+#include "../tm_examples.h"
+
+    int tm_example_vww();
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*TM_EXAMPLE_VWW_H*/
