+++ include/tm_port.h
@@ -26,7 +26,11 @@ limitations under the License.
 
 /******************************* PORT CONFIG  ************************************/
 #define TM_ARCH         TM_ARCH_CPU
+
+#ifndef TM_OPT_LEVEL
 #define TM_OPT_LEVEL    TM_OPT0
+#endif
+
 #define TM_MDL_TYPE     TM_MDL_INT8
 #define TM_FASTSCALE    (0)         //enable if your chip don't have FPU, may speed up 1/3, but decrease accuracy
 #define TM_LOCAL_MATH   (0)         //use local math func (like exp()) to avoid libm
