+++ tools/requirements.txt
@@ -1,0 +1,3 @@
+tensorflow
+numpy
+Pillow
