+++ tools/tmdl/cifar10_f.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #define MDL_BUF_LEN (45056)
 #define LBUF_LEN (205200)
-const uint8_t mdl_data[358776]={\
+static const uint8_t mdl_data[358776]={\
 	0x4d, 0x41, 0x49, 0x58, 0x02, 0x01, 0x01, 0x00, 0x01, 0x00, 0x06, 0x00, 0x00, 0xb0, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x20, 0x00, 0x20, 0x00, 0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -22428,6 +22428,7 @@ const uint8_t mdl_data[358776]={\
 	0x32, 0x92, 0x75, 0xbe, 0x75, 0xcb, 0xf2, 0xbe, 0x03, 0x00, 0x01, 0x00, 0x30, 0x00, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0xd8, 0xaf, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x0a, 0x00, 
 	0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x0a, 0x00, 0x00, 0x00, 0x80, 0x3f, 0x00, 0x00, 0x00, 0x00, 
-	0x00, 0x00, 0x80, 0x3f, 0x00, 0x00, 0x00, 0x00, };
+	0x00, 0x00, 0x80, 0x3f, 0x00, 0x00, 0x00, 0x00, 
+};
 
 #endif
