+++ tools/tmdl/cifar10_q.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #define MDL_BUF_LEN (11264)
 #define LBUF_LEN (51856)
-const uint8_t mdl_data[91008]={\
+static const uint8_t mdl_data[91008]={\
 	0x4d, 0x41, 0x49, 0x58, 0x00, 0x01, 0x01, 0x00, 0x01, 0x00, 0x06, 0x00, 0x00, 0x2c, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x20, 0x00, 0x20, 0x00, 0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -5693,6 +5693,7 @@ const uint8_t mdl_data[91008]={\
 	0x03, 0x00, 0x01, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xc8, 0x2b, 0x00, 0x00, 
 	0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x0a, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x0a, 0x00, 
 	0xb7, 0x7b, 0x8d, 0x3d, 0x37, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x3b, 0x80, 0xff, 0xff, 0xff, 
-	};
+	
+};
 
 #endif
