+++ tools/tmdl/mbnet128_0.25_q.h
@@ -1,10 +1,10 @@
-#ifndef __MODEL_FILE__H
-#define __MODEL_FILE__H
+#ifndef MBNET128_0_25_Q_H
+#define MBNET128_0_25_Q_H
 
 #include <stdint.h>
-#define MDL_BUF_LEN (98304)
-#define LBUF_LEN (264144)
-const uint8_t mdl_data[495936]={\
+#define MBNET128_0_25_Q_MDLBUFLEN (98304)
+#define MBNET128_0_25_Q_LBUFLEN (264144)
+static const uint8_t mdl_data[495936]={\
 	0x4d, 0x41, 0x49, 0x58, 0x00, 0x01, 0x01, 0x00, 0x01, 0x00, 0x1f, 0x00, 0x00, 0x80, 0x01, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x80, 0x00, 0x80, 0x00, 0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0xe8, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
