+++ tools/tmdl/mbnet96_0.25_q.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #define MDL_BUF_LEN (55296)
 #define LBUF_LEN (264144)
-const uint8_t mdl_data[495984]={\
+static const uint8_t mdl_data[495984]={\
 	0x4d, 0x41, 0x49, 0x58, 0x00, 0x01, 0x01, 0x00, 0x01, 0x00, 0x20, 0x00, 0x00, 0xd8, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x60, 0x00, 0x60, 0x00, 0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0xe8, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -31004,6 +31004,7 @@ const uint8_t mdl_data[495984]={\
 	0x04, 0x00, 0x01, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
 	0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 0xe8, 0x03, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0xe8, 0x03, 
 	0x00, 0x00, 0x80, 0x3b, 0x80, 0xff, 0xff, 0xff, 0x00, 0x00, 0x80, 0x3b, 0x80, 0xff, 0xff, 0xff, 
-	};
+	
+};
 
 #endif
