+++ tools/tmdl/mnist_fp8_143.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #define MDL_BUF_LEN (3528)
 #define LBUF_LEN (73968)
-const uint8_t mdl_data[88504]={\
+static const uint8_t mdl_data[88504]={\
 	0x4d, 0x41, 0x49, 0x58, 0x04, 0x01, 0x01, 0x00, 0x01, 0x00, 0x07, 0x00, 0xc8, 0x0d, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x1c, 0x00, 0x1c, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -5533,4 +5533,5 @@ const uint8_t mdl_data[88504]={\
 	0x29, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x01, 0x00, 0x30, 0x00, 0x00, 0x00, 
 	0xb8, 0x0d, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x0a, 0x00, 
 	0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x0a, 0x00, 0x00, 0x00, 0x80, 0x3f, 0x00, 0x00, 0x00, 0x00, 
-	0x00, 0x00, 0x80, 0x3f, 0x00, 0x00, 0x00, 0x00, };
+	0x00, 0x00, 0x80, 0x3f, 0x00, 0x00, 0x00, 0x00, 
+};
