+++ tools/tmdl/mnist_valid_f.h
@@ -1,10 +1,10 @@
-#ifndef __MODEL_FILE__H
-#define __MODEL_FILE__H
+#ifndef MNIST_VALID_F_H
+#define MNIST_VALID_F_H
 
 #include <stdint.h>
-#define MDL_BUF_LEN (5840)
-#define LBUF_LEN (4816)
-const uint8_t mdl_data[7160]={\
+#define MNIST_VALID_F_MDLBUFLEN (5840)
+#define MNIST_VALID_F_LBUFLEN (4816)
+static const uint8_t mdl_data[7160]={\
 	0x4d, 0x41, 0x49, 0x58, 0x02, 0x01, 0x01, 0x00, 0x01, 0x00, 0x06, 0x00, 0xd0, 0x16, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x1c, 0x00, 0x1c, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
