+++ tools/tmdl/mnist_valid_q.h
@@ -1,10 +1,10 @@
-#ifndef __MODEL_FILE__H
-#define __MODEL_FILE__H
+#ifndef MNIST_VALID_Q_H
+#define MNIST_VALID_Q_H
 
 #include <stdint.h>
-#define MDL_BUF_LEN (1464)
-#define LBUF_LEN (1424)
-const uint8_t mdl_data[2408]={\
+#define MNIST_VALID_Q_MDLBUFLEN (1464)
+#define MNIST_VALID_Q_LBUFLEN (1424)
+static const uint8_t mdl_data[2408]={\
 	0x4d, 0x41, 0x49, 0x58, 0x00, 0x01, 0x01, 0x00, 0x01, 0x00, 0x06, 0x00, 0xb8, 0x05, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x1c, 0x00, 0x1c, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0x0a, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
