+++ tools/tmdl/vww96_q.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #define MDL_BUF_LEN (55296)
 #define LBUF_LEN (67728)
-const uint8_t mdl_data[232416]={\
+static const uint8_t mdl_data[232416]={\
 	0x4d, 0x41, 0x49, 0x58, 0x00, 0x01, 0x01, 0x00, 0x01, 0x00, 0x1e, 0x00, 0x00, 0xd8, 0x00, 0x00, 
 	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x60, 0x00, 0x60, 0x00, 0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 
 	0x01, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -14531,6 +14531,7 @@ const uint8_t mdl_data[232416]={\
 	0x03, 0x00, 0x01, 0x00, 0x30, 0x00, 0x00, 0x00, 0xf8, 0xd7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
 	0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x02, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x02, 0x00, 
 	0xd2, 0x98, 0xa2, 0x3c, 0x80, 0xff, 0xff, 0xff, 0x00, 0x00, 0x80, 0x3b, 0x80, 0xff, 0xff, 0xff, 
-	};
+	
+};
 
 #endif
